Disable hardware accellerated drawing.
It doesn't work correctly for the cursor. Instead of XOR-ing over
the text it draws opaque, so you can't see the character under the
cursor.

We need to figure out a different way of rendering the cursor
that is compatible with hardware acceleration.

Added hardware acceleration drawing to preferences pane, that works OK.