Update AndroidManifest.xml
These additions allow multi-window devices to resize ATE so that it can be displayed with one other application. These windows are resizable by the user and simply touching one or the other changes the focus. The effect, with ATE as one of the applications, can be that of an IDE if the other application is a script editor. One such multi-window device is the Galaxy Note 3.