Add "Term Here" send intent to open a terminal in a given directory.
This addition adds "Term here" to the "Send" or "Send to" option on many file managers,
accessable by a long press. Long-press a file or directory in a file manager and select
"Term here"; Term will open to a new window at that location.

The permission.RUN_SCRIPT permission is not used because no executables other than 'cd'
are invoked. Choosing "Term here" performs essentially the same action as merely opening
Term, but with the addition of the specified starting directory.

I chose to use the "send" intent option instead of a normal intent with 'data' because
I think showing "Term here" for every file you wish to open normally is excessive, so
for the special case where you want to open a terminal at a file or directory location
the "send" feature can be used.

It might be desired to later add a the ability to launch scripts from a file manager
by selecting the script, or other executable, and choosing "Term", or "Run with Term".
This would be implemented with a normal Intent which uses 'data' to communicate the file
and would be selected after a short-press in most or all file managers.

Original patch by Frank Westlake. Improved formatting, error checking,
and shell quoting by jackpal.

Tested in "ES File Explorer File Manager".